LayoutTests/fast/dom/domListEnumeration-expected.txt

This tests enumerating the elements of DOM lists.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



[object NodeList]
PASS resultArray.length is 5
PASS resultArray[0].i is '0'
PASS resultArray[0].item is nodeList.item(0)
PASS resultArray[1].i is '1'
PASS resultArray[1].item is nodeList.item(1)
PASS resultArray[2].i is '2'
PASS resultArray[2].item is nodeList.item(2)

[object HTMLCollection]
PASS resultArray.length is 7
PASS resultArray[0].i is '0'
PASS resultArray[0].item is htmlCollection.item(0)
PASS resultArray[1].i is '1'
PASS resultArray[1].item is htmlCollection.item(1)
PASS resultArray[2].i is '2'
PASS resultArray[2].item is htmlCollection.item(2)

[object NamedNodeMap]
PASS resultArray.length is 11
PASS resultArray[0].i is '0'
PASS resultArray[0].item is namedNodeMap.item(0)
PASS resultArray[1].i is '1'
PASS resultArray[1].item is namedNodeMap.item(1)
PASS resultArray[2].i is '2'
PASS resultArray[2].item is namedNodeMap.item(2)

[object HTMLFormElement]
PASS resultArray.length is 134
PASS resultArray[0].i is '0'
PASS resultArray[0].item is document.getElementsByTagName('select')[0]
PASS resultArray[1].i is '1'
PASS resultArray[1].item is document.getElementsByTagName('select')[1]
PASS resultArray[2].i is '2'
PASS resultArray[2].item is document.getElementsByTagName('select')[2]

[object HTMLSelectElement]
PASS resultArray.length is 140
PASS resultArray[0].i is '0'
PASS resultArray[0].item is document.getElementsByTagName('option')[0]
PASS resultArray[1].i is '1'
PASS resultArray[1].item is document.getElementsByTagName('option')[1]
PASS resultArray[2].i is '2'
PASS resultArray[2].item is document.getElementsByTagName('option')[2]

[object StyleSheetList]
PASS resultArray.length is 5
PASS resultArray[0].i is '0'
PASS resultArray[0].item is styleSheetList.item(0)
PASS resultArray[1].i is '1'
PASS resultArray[1].item is styleSheetList.item(1)
PASS resultArray[2].i is '2'
PASS resultArray[2].item is styleSheetList.item(2)

[object CSSRuleList]
PASS resultArray.length is 5
PASS resultArray[0].i is '0'
PASS resultArray[0].item is cssRuleList.item(0)
PASS resultArray[1].i is '1'
PASS resultArray[1].item is cssRuleList.item(1)
PASS resultArray[2].i is '2'
PASS resultArray[2].item is cssRuleList.item(2)

[object CSSStyleDeclaration]
PASS resultArray.length is 14
PASS resultArray[0].i is '0'
PASS resultArray[0].item is cssStyleDeclaration.item(0)
PASS resultArray[1].i is '1'
PASS resultArray[1].item is cssStyleDeclaration.item(1)
PASS resultArray[2].i is '2'
PASS resultArray[2].item is cssStyleDeclaration.item(2)

[object CSSValueList]
PASS resultArray.length is 10
PASS resultArray[0].i is '0'
PASS resultArray[0].item is cssValueList.item(0)
PASS resultArray[1].i is '1'
PASS resultArray[1].item is cssValueList.item(1)

[object MediaList]
PASS resultArray.length is 8
PASS resultArray[0].i is '0'
PASS resultArray[0].item is mediaList.item(0)
PASS resultArray[1].i is '1'
PASS resultArray[1].item is mediaList.item(1)
PASS resultArray[2].i is '2'
PASS resultArray[2].item is mediaList.item(2)

PASS successfullyParsed is true

TEST COMPLETE
